requirements.txt
tensorflow==2.15.0
numpy
pillow
scikit-learn
tqdm
fastapi
uvicorn
